modified:   app/Controllers/AjaxController.php

+++ app/Controllers/AjaxController.php
@@ -664,10 +664,10 @@ class AjaxController extends BaseController {
 																	</div>
 																</td>';
 																
-												 $basic_won  =  $room['goods_price1'] * $room['baht_thai'];
+												 $basic_won  =  (int)($room['goods_price1'] * $room['baht_thai']);
 												 $basic_bath =  $room['goods_price1'];
 											   
-												 $price_won  = ($room['goods_price2'] + $room['goods_price3']) * $room['baht_thai'];
+												 $price_won  =  (int)(($room['goods_price2'] + $room['goods_price3']) * $room['baht_thai']);
 												 $price_bath =  $room['goods_price2'] + $room['goods_price3'];
 																
 												 $msg .= '<td>
